move cache stack to advance stack
